Fix processing some YAML files without START_OF_DOCUMENT
The parser relies on the `\n` that precedes most items to process the
documents.  A YAML file without the `---\n` prevents the system to be
able to detect the first item, leading to inconsistent parsing.

Automatically prepend a start of document if none is found when reading
the YAML file, and mandate the presence of the START_OF_DOCUMENT token
in the parser.

Fixes #3