Add support for &anchor / *alias
YAML provides node anchors (using `&`) and aliases (using `*`) to reuse
portion of a document.

When reading an anchor, store the actual data and act like an alias was
found.  When writing the document, output the actual data as an anchor
when the first alias is found, and output on alias for the next
occurences.